Switched to Dictionary based AssetRepository
It helps with keeping track of coordinates of each asset on a playBoard.
I also moved check if Daniel has been cought to newly created DanielController.
Also created TreeController and I have tweaked a little how the trees appear on the playBoard.
Every asset that is placed on the playBoard is also added to the AssetRepository.
Same with deleting. Its simoultonus now to keep consistancy between AssetRepository and playBoard.
+ Minor changes to assure that.